requirements.txt
# Core dependencies
python-dotenv>=1.0.0
httpx>=0.27.0

# LangChain ecosystem
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0

# Search
tavily-python>=0.4.0

# MCP (Model Context Protocol)
mcp>=1.0.0

# Utilities
rich>=13.0.0

# Developer UI
streamlit>=1.37.0

# Document processing
pypdf>=4.0.0
python-docx>=1.0.0
pdfplumber>=0.10.0
pdf2image>=1.16.0
requests>=2.28.0
